+++ src/Gyselroth/Helper/Interfaces/ConstantsCountryCodes.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Copyright (c) 2017-2020 gyselroth™  (http://www.gyselroth.net)
+ *
+ * @package \gyselroth\Helper
+ * @author  gyselroth™  (http://www.gyselroth.com)
+ * @link    http://www.gyselroth.com
+ * @license Apache-2.0
+ */
+
+namespace Gyselroth\Helper\Interfaces;
+
+interface ConstantsCountryCodes
+{
+    public const COUNTRY_CODE_SWITZERLAND   = 'ch';
+    public const COUNTRY_CODE_GERMANY       = 'de';
+    public const COUNTRY_CODE_UNITED_STATES = 'us';
+}